Move resolver mod under kuadrant
