+++ src/v2/kuadrant/mod.rs
@@ -1,5 +1,9 @@
 pub mod cache;
 pub mod context;
+mod resolver;
+
+#[cfg(test)]
+pub use resolver::MockWasmHost;
 
 pub use cache::AttributeCache;
 pub use context::ReqRespCtx;
