+++ src/v2/kuadrant/resolver/wasm_host.rs
