+++ src/v2/mod.rs
@@ -1,7 +1,4 @@
 #[allow(dead_code)]
-pub mod resolver;
-
-#[allow(dead_code)]
 pub mod kuadrant;
 
 #[allow(dead_code)]
